drift-detection/pyproject.toml
[project]
name = "attest-example-drift-detection"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["attest-ai>=0.4.1", "pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
testpaths = ["."]
asyncio_mode = "auto"
